docs: replace rule(s) with validation(s)
