--- apps/docs-app/src/app/docs/02-configuring-rules/ng-doc.category.ts
@@ -1,10 +1,0 @@
-import { NgDocCategory } from '@ng-doc/core';
-
-const ConfiguringRules: NgDocCategory = {
-  title: 'Configuring Rules',
-  expandable: true,
-  route: 'configuring-rules',
-  order: 2
-};
-
-export default ConfiguringRules;
