--- apps/docs-app/src/app/docs/02-configuring-rules/overriding-the-property-name/ng-doc.page.ts
@@ -1,12 +1,0 @@
-import {NgDocPage} from '@ng-doc/core';
-import ConfiguringRules from '../ng-doc.category';
-
-const OverridingThePropertyName: NgDocPage = {
-	title: `Overriding the Property Name`,
-	mdFile: './index.md',
-	route: `overriding-the-property-name`,
-	category: ConfiguringRules,
-	order: 2,
-};
-
-export default OverridingThePropertyName;
