+++ apps/docs-app/src/app/docs/02-configuring-validations/ng-doc.category.ts
@@ -1,0 +1,10 @@
+import { NgDocCategory } from '@ng-doc/core';
+
+const ConfiguringValidations: NgDocCategory = {
+  title: 'Configuring Validations',
+  expandable: true,
+  route: 'configuring-validations',
+  order: 2
+};
+
+export default ConfiguringValidations;
