+++ apps/docs-app/src/app/docs/02-configuring-validations/override-message/ng-doc.page.ts
@@ -1,11 +1,11 @@
 import { NgDocPage } from '@ng-doc/core';
-import ConfiguringRules from '../ng-doc.category';
+import ConfiguringValidations from '../ng-doc.category';
 
 const OverrideMessage: NgDocPage = {
   title: `Overriding the Message`,
   mdFile: './index.md',
   route: `override-message`,
-  category: ConfiguringRules,
+  category: ConfiguringValidations,
   order: 1
 };
