+++ apps/docs-app/src/app/docs/02-configuring-validations/overriding-the-property-name/ng-doc.page.ts
@@ -1,0 +1,12 @@
+import { NgDocPage } from '@ng-doc/core';
+import ConfiguringValidations from '../ng-doc.category';
+
+const OverridingThePropertyName: NgDocPage = {
+  title: `Overriding the Property Name`,
+  mdFile: './index.md',
+  route: `overriding-the-property-name`,
+  category: ConfiguringValidations,
+  order: 2
+};
+
+export default OverridingThePropertyName;
