+++ apps/docs-app/src/app/docs/03-building-validations/builtin-validations/index.md
@@ -1,5 +1,5 @@
 ---
-keyword: 'BuildingRulesBuiltInRules'
+keyword: 'BuildingValidationsBuiltInValidations'
 ---
 
 {% import "../../templates/macros/validations.md" as validations %}
